% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_and_outlier_recognition.R
\name{clustering_and_outlier_recognition}
\alias{clustering_and_outlier_recognition}
\title{Description of two_layer_partition_clustering.R.
This function implements the two-layer partition clustering.
See Robust two-layer partition clustering of sparse multivariate functional data,
Zhuo Qu, Wenlin Dai, and Marc G. Genton, https://www.sciencedirect.com/science/article/pii/S2452306223000138
This function implements algorithm 3 in the paper.}
\usage{
clustering_and_outlier_recognition(
  cluster_result,
  minprop = 0.05,
  distance_matrix,
  alpha_c = 0.85
)
}
\arguments{
\item{cluster_result}{is the list of the clustering function. Each list includes center, and element}

\item{minprop}{minimal required number of elements in a cluster / the number of samples in the population.
By default, minprop = 0.05.}

\item{distance_matrix}{is the matrix of distances for the population.}

\item{alpha_c}{is used to defines the alpha_c quantile of the distances. By default, alpha_c = 0.85.
between curves inside a cluster and the cluster center.}
}
\value{
A list of three elements:
1. clust is the result of the robust two-layer partition clustering, where each sublist is a subclust including center and element.
2. isolate is the result of outliers in the robust two-layer partition clustering.
3. criteria_matrix is a matrix of the distance between each potential outlier and the center of all clust. The cut-off values are the distance
threshold determined by alpha_c.
}
\description{
Description of two_layer_partition_clustering.R.
This function implements the two-layer partition clustering.
See Robust two-layer partition clustering of sparse multivariate functional data,
Zhuo Qu, Wenlin Dai, and Marc G. Genton, https://www.sciencedirect.com/science/article/pii/S2452306223000138
This function implements algorithm 3 in the paper.
}
\examples{
library(fda); canada_temp <- t(CanadianWeather$monthlyTemp) ## a matrix of 35 stations and 12 months
 n <- nrow(canada_temp); nt <- ncol(canada_temp); time <- 1:12
 complete_canada.temp_list <- lapply(1:n, function(k) {
 res <- list(argvals = time,
             subj = rep(k, 12), y = canada_temp[k, ])
  return(res)
  })
  dist_complete_canada.temp <- elastic_time_distance(complete_canada.temp_list)
  neighbors_canada.temp <- findneighbours(dist_complete_canada.temp)
 no.of.neighbors <- apply(neighbors_canada.temp, 1, sum)
 two_layer_partition <- two_layer_partition_clustering(neighbors_canada.temp)
 cluster_canada.temp <- two_layer_partition[[2]]
 clust_outlier_canada.temp <- clustering_and_outlier_recognition(cluster_canada.temp, minprop = 0.1, distance_matrix = dist_complete_canada.temp)
}
