% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_outlying.R
\name{directional_outlying}
\alias{directional_outlying}
\title{Description of directional_outlying.R.
Obtain the directional outlyingness for univariate functional data or multivariate functional data}
\usage{
directional_outlying(
  data,
  diroutmatrix = FALSE,
  depth.dir = "RP",
  d_value = TRUE,
  sq_vo = FALSE
)
}
\arguments{
\item{data}{is either a list of length = number of variables, and each sublist is a nobs * length(timegrid) matrix,
or data is a matrix of nobs rows and length(timegrid) columns.}

\item{diroutmatrix}{is a logical variable of whether to obtain the directional outlyingness. By default,
diroutmatrix = FALSE}

\item{depth.dir}{is a chracter variable in c("RP","MhD","SD","HS"), where "RP" represents the random
projection depth, "MhD" represents the Mahalanobis depth, "SD" represents the simplicial depth, and
"HS" represents the random halfspace depth. By default, depth.dir = "RP".}

\item{d_value}{is a logical variable of whether to return the squared Mahalanobis distance}

\item{sq_vo}{is a logical variable of whether to use the variational outlyingness or the
square root of the variational outlyingness. By default, sq_vo is FALSE.}
}
\value{
a list including the index of the mean outlyingness and variational outlyingness
}
\description{
Description of directional_outlying.R.
Obtain the directional outlyingness for univariate functional data or multivariate functional data
}
\examples{
library(spam);n = 200;p = 4;Sigmainv <- .25^abs(outer(1:n,1:n,"-"))
Sigmainv <- as.spam( Sigmainv, eps=1e-4)
Sigma <- solve( Sigmainv);data <- rmvnorm(n, mu = rep(0, p), Sigma = Sigma)
directional_outlying(data, diroutmatrix = FALSE, depth.dir = "RP", d_value = TRUE, sq_vo = FALSE)
}
