% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faccal_asy.R
\name{faccal_asy}
\alias{faccal_asy}
\title{Description of faccal_asy.R.
to estimate the the factor and the cutoff value of the approximate F distribution of factor times the distance matrix
 between any observation and minimum determinant covariance mean.
See the Appendix in The Distribution of Robust Distances (2005), Johanna Hardin & David M Rocke,
Journal of Computational and Graphical Statistics,14(4), Pages 928--946
 https://www.tandfonline.com/doi/epdf/10.1198/106186005X77685?needAccess=true}
\usage{
faccal_asy(n, dim)
}
\arguments{
\item{n}{is the number of observations in the multivariate data}

\item{dim}{is the dimension of the data}
}
\value{
a list showing the factor and the cutoff value
}
\description{
Description of faccal_asy.R.
to estimate the the factor and the cutoff value of the approximate F distribution of factor times the distance matrix
 between any observation and minimum determinant covariance mean.
See the Appendix in The Distribution of Robust Distances (2005), Johanna Hardin & David M Rocke,
Journal of Computational and Graphical Statistics,14(4), Pages 928--946
 https://www.tandfonline.com/doi/epdf/10.1198/106186005X77685?needAccess=true
}
\examples{
faccal_asy(100, 3)
}
