% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faccal_num.R
\name{faccal_num}
\alias{faccal_num}
\title{Description of faccal_num.R.
to estimate the the factor and the cutoff value of the approximate F distribution of factor times the distance matrix
 between any observation and minimum determinant covariance mean.
 This function takes the number of observations and the number of variables into account.
See The Distribution of Robust Distances (2005), Johanna Hardin & David M Rocke,
Journal of Computational and Graphical Statistics,14(4), Pages 928--946
 https://www.tandfonline.com/doi/epdf/10.1198/106186005X77685?needAccess=true}
\usage{
faccal_num(n, dim)
}
\arguments{
\item{n}{is the number of observations in the multivariate data}

\item{dim}{is the dimension of the data}
}
\value{
a list showing the factor and the cutoff value
}
\description{
Description of faccal_num.R.
to estimate the the factor and the cutoff value of the approximate F distribution of factor times the distance matrix
 between any observation and minimum determinant covariance mean.
 This function takes the number of observations and the number of variables into account.
See The Distribution of Robust Distances (2005), Johanna Hardin & David M Rocke,
Journal of Computational and Graphical Statistics,14(4), Pages 928--946
 https://www.tandfonline.com/doi/epdf/10.1198/106186005X77685?needAccess=true
}
\examples{
faccal_num(100, 3)
}
