% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findneighbours.R
\name{findneighbours}
\alias{findneighbours}
\title{
This function returns a logical matrix. Row i shows whether each subject is a neighbor of subject i.}
\usage{
findneighbours(distance_matrix, theta = 0.05)
}
\arguments{
\item{distance_matrix}{is the matrix of distances for the population.}

\item{theta}{is the probability (0 < theta < 1) for the cut-off quantile. By default, theta = 0.05.}
}
\value{
a logical matrix
}
\description{
Description of findneighbours.R.
This function returns a logical matrix. Each row shows whether each subject is the neighbor.
}
\examples{
library(fda); canada_temp <- t(CanadianWeather$monthlyTemp) ## a matrix of 35 stations and 12 months
n <- nrow(canada_temp); nt <- ncol(canada_temp); time <- 1:12
complete_canada.temp_list <- lapply(1:n, function(k) {
res <- list(argvals = time,
            subj = rep(k, 12), y = canada_temp[k, ])
 return(res)
 })
 dist_complete_canada.temp <- elastic_time_distance(complete_canada.temp_list)
 neighbors_canada.temp <- findneighbours(dist_complete_canada.temp)
no.of.neighbors <- apply(neighbors_canada.temp, 1, sum)
}
