% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplot_modified.R
\name{msplot_modified}
\alias{msplot_modified}
\title{Description of msplot_modified.R.
Obtain the Magnitude-Shape plot}
\usage{
msplot_modified(
  data,
  depth.dir = "RP",
  plot = FALSE,
  col.out = "green",
  col.med = "purple",
  col.normal = "black",
  plot.type = "scatter",
  sq_vo = FALSE,
  datatopic = NULL
)
}
\arguments{
\item{data}{is a list with d (the number of variables). And each list is a n*time_nb matrix, where n is the number of
observations, and time_nb is the number of the time grid.}

\item{depth.dir}{is a chracter variable in c("RP","MhD","SD","HS"), where "RP" represents the random
projection depth, "MhD" represents the Mahalanobis depth, "SD" represents the simplicial depth, and
"HS" represents the random halfspace depth. By default, depth.dir = "RP".}

\item{plot}{is a logical variable of whether to show a figure.}

\item{col.out}{is the color of outliers. By default, col.out = "green".}

\item{col.med}{is the color of the median. By default, col.med = "purple".}

\item{col.normal}{is the color of the non-outlyinge curves. By default, col.normal = "black".}

\item{plot.type}{is either "parallel" or "scatter". By default, plot.type = "scatter".}

\item{sq_vo}{is a logical variable of whether to use the variational outlyingness or the
square root of the variational outlyingness. By default, sq_vo is FALSE.}

\item{datatopic}{is the description of your data. This will be shown in the title of the figure.}
}
\value{
a list of 4 elements:
1. mo. If data is a list, then mo is a matrix of n * p, where n is the number of samples,
and p is the number of variables.
2. vo. vo is a vector of n, and n is the number of samples in the population.
3. the index of outliers from the directional outlyingness
4. the index of the median curve from the directional outlyingness.
}
\description{
Description of msplot_modified.R.
Obtain the Magnitude-Shape plot
}
\examples{
library(fda)
library(roahd)
canada_temp <- t(CanadianWeather$monthlyTemp)
canada_precip <- t(CanadianWeather$monthlyPrecip)
msplot_modified(data = canada_temp,
             plot = TRUE, sq_vo = FALSE)
msplot_modified(data = list(temp = canada_temp, precip = canada_precip),
             plot = TRUE, sq_vo = FALSE)

}
