% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_dirout.R
\name{outlier_dirout}
\alias{outlier_dirout}
\title{Description of outlier_dirout.R.
Obtain the index of outlier, index of median curve and the measure of outlyingness
with directional outlyingness method
for univariate functional data or multivariate functional data.}
\usage{
outlier_dirout(data, sq_vo = FALSE, depth.dir = "RP")
}
\arguments{
\item{data}{is either a list of length p, and each sublist is a nobs * length(timegrid) matrix,
or data is a matrix of nobs rows and length(timegrid) columns.}

\item{sq_vo}{is a logical variable of whether to use the variational outlyingness or the
square root of the variational outlyingness. By default, sq_vo is FALSE.}

\item{depth.dir}{is a chracter variable in c("RP","MhD","SD","HS"), where "RP" represents the random
projection depth, "MhD" represents the Mahalanobis depth, "SD" represents the simplicial depth, and
"HS" represents the random halfspace depth. By default, depth.dir = "RP".}
}
\value{
a list including the index of the mean outlyingness and variational outlyingness
}
\description{
Description of outlier_dirout.R.
Obtain the index of outlier, index of median curve and the measure of outlyingness
with directional outlyingness method
for univariate functional data or multivariate functional data.
}
\examples{
library(spam);n = 200;p = 4;Sigmainv <- .25^abs(outer(1:p,1:p,"-"))
Sigmainv <- as.spam( Sigmainv, eps=1e-4)
Sigma <- solve( Sigmainv);data <- rmvnorm(n, mu = rep(0, p), Sigma = Sigma)
outlier_dirout(data, sq_vo = FALSE)
}
