% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlying_depth.R
\name{outlying_depth}
\alias{outlying_depth}
\title{Description of outlying_depth.R.
Obtain the directional outlyingness for multivariate data}
\usage{
outlying_depth(data_mat, depth.dir)
}
\arguments{
\item{data_mat}{a matrix with no. rows = number of observations, and no. columns = number of variables.}

\item{depth.dir}{chracter variable in "RP" (Random Projection depth), "MhD" (Mahalanobis depth),
"SD" (Simplicial depth), "HS" (Random Halfspace depth).}
}
\value{
a vector with length = number of observations. It measures the outlyingness of the data which is an inverse of the depth.
}
\description{
Description of outlying_depth.R.
Obtain the directional outlyingness for multivariate data
}
\examples{
library(spam); n = 200; p <- 4; Sigmainv <- .25^abs(outer(1:p,1:p,"-"))
Sigmainv <- as.spam( Sigmainv, eps=1e-4)
Sigma <- solve(Sigmainv)
data <- rmvnorm(n, mean = rep(0, p), Sigma = Sigma)
outlying_depth(data, "RP")
}
