% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetlist.R
\name{subsetlist}
\alias{subsetlist}
\title{Description of subsetlist.R.
More information about what subsetlist.R function does.}
\usage{
subsetlist(myList, elementNames)
}
\arguments{
\item{myList}{Input parameter is a list}

\item{elementNames}{Input parameter is the element index that of interests.}
}
\value{
a subset list
}
\description{
Description of subsetlist.R.
More information about what subsetlist.R function does.
}
\examples{
subsetlist(list(l1 = c(1, 2, 3), l2 = c("d", "efd", "g"), l3 =c(12,345), l4 = 1), c(1, 3))
}
